requirements.txt
streamlit==1.32.2
python-dotenv==1.0.1

langchain==0.2.14
langchain-community==0.2.12
langchain-core==0.2.32
langchain-text-splitters==0.2.2
langchain-chroma==0.1.2
langchain-huggingface==0.0.3
langchain-groq==0.1.5

sentence-transformers==2.7.0
pymupdf==1.24.1